sentinal-2.0.0 (#32)
* sqlite3 conversion, initial commit

* cleanup var types

* cleanup pcrefind

* misplaced time()

* dircount unneeded

* cleanup main loop

* don't count dirbytes when dirlimit is not set

* bug fix, stifle dryrun output

* some rewrites, some cleanup

* add drcount to process_dirs()

* misc cleanup

* more int-to-short conversions

* more types cleanup

* add var comments

* add var comments

* add long_options

* more var comments

* increase padding

* fix failure return value

* expire and rmdir need more reporting

* s/expired/removed/

* dfs needs more reporting

* fullpaths in debian.sh

+++ logname.c
@@ -31,7 +31,7 @@ static void substrval(char *, char *, time_t);
 char   *logname(char *template, char *filename)
 {
 	struct tm tbuf;
-	time_t  curtime;
+	time_t  curtime;								/* now */
 
 	/*
 	 * filename becomes the completed template
@@ -65,7 +65,7 @@ static void substrval(char *template, char *token, time_t value)
 	char    search[PATH_MAX];
 	char    valbuf[PATH_MAX];
 	char   *p;
-	int     len;
+	size_t  len;
 
 	if(IS_NULL(token) || *token != '%')
 		return;
